requirements.txt
PyQT5~=5.15.0
rethinkdb~=2.4.7
qrcode~=6.1
pillow
reportlab~=3.5.34
colorama~=0.4.3
Flask~=1.1.2
argon2~=0.1.10
python-dateutil~=2.7.3
setuptools~=45.2.0
africastalking~=1.2.3
pytest~=6.1.1
bcrypt~=3.1.7
itsdangerous~=1.1.0
requests~=2.22.0
html-testRunner~=1.2.1
dotenv